Added LoadingTable component to show a loading screen for tables. Added a loading table on ratings page
