--- src/Components/drawer/problemSet/problemsList.js
@@ -1,126 +1,0 @@
-const problemsList = [
-  {
-    contestId: 'GBYE19',
-    problemId: 'GBYE19G',
-    problemName: 'The Chat pattern',
-    tags: ['adhoc', 'strings'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'GBYE19',
-    problemId: 'GBYE19F',
-    problemName: 'TML events',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'GBYE19',
-    problemId: 'GBYE19A',
-    problemName: 'Swedish House Mafia',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'SRM05',
-    problemId: 'SRM05A',
-    problemName: 'Cibin Likes Networking',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'wrong',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18B',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'accepted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18C',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18D',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'strings'],
-    points: '100',
-    submissionStatus: 'wrong',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18E',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['dp', 'math'],
-    points: '100',
-    submissionStatus: 'accepted',
-  },
-
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18F',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'dp'],
-    points: '100',
-    submissionStatus: 'wrong',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18G',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'accepted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18H',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'accepted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18I',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'accepted',
-  },
-
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18J',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18K',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-  {
-    contestId: 'SEPT18',
-    problemId: 'SEPT18L',
-    problemName: 'Shweta And Her Rakhi',
-    tags: ['adhoc', 'math'],
-    points: '100',
-    submissionStatus: 'unsubmitted',
-  },
-];
-
-export default problemsList;
