+++ src/Components/drawer/ratings/Ratings.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Headline4, Body2 } from '@material/react-typography';
+import { Body2 } from '@material/react-typography';
 import PropTypes from 'prop-types';
 import RatingsTable from './RatingsTable';
 import 'tachyons';
@@ -8,7 +8,6 @@ const Ratings = ({
   users, activePageNumber, limit, ratingsUpdatedTill,
 }) => (
   <div>
-    <Headline4 className="purple mb2">Ratings</Headline4>
     <Body2 className="mid-gray ml1 mt2">
       Ratings have been updated till
       &nbsp;
