--- src/Components/drawer/ratings/ratings.js
@@ -1,90 +1,0 @@
-const ratings = [
-  {
-    rank: 1,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 2,
-    name: 'Ashok Choudhary',
-    profileId: '5b5d756b37392f89933e23124124',
-    rating: 1000,
-  },
-  {
-    rank: 3,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-
-  {
-    rank: 4,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-
-  {
-    rank: 5,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 6,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 7,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 8,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 9,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 10,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 11,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 12,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 13,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-  {
-    rank: 14,
-    name: 'Vipul Singh Raghuvanshi',
-    profileId: '5b5d756b37392f89933e7515',
-    rating: 1566,
-  },
-];
-
-export default ratings;
